% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asl}
\alias{asl}
\title{Age, sex and length of Chinook salmon}
\format{
A data frame with 49,369 rows and 4 variables:
\describe{
\item{year}{Year the fish was sampled.}
\item{sex}{Sex of the fish (factor: \code{Female}, \code{Male}).}
\item{age}{Age of the fish (years).}
\item{length}{Mid-eye to fork length (mm).}
}
}
\source{
Simulated by B. Staton,
\url{https://github.com/bstaton1/au-r-workshop-data}.
}
\usage{
asl
}
\description{
Ages, sexes and lengths of about 50,000 Chinook salmon
(\emph{Oncorhynchus tshawytscha}) sampled over 50 years (1966--2015).
}
\keyword{datasets}
