% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ButterflyCounts}
\alias{ButterflyCounts}
\title{ButterflyCounts}
\format{
A data frame with 118 rows and 4 variables:
\describe{
\item{\code{year}}{integer year of data collection}
\item{\code{date}}{character date of data collection}
\item{\code{count}}{integer count of butterflies}
\item{\code{DOY}}{double day of year associated with the count}
}
}
\source{
Edwards, C. B., & Crone, E. E. (2021). Estimating abundance and phenology from transect count data with GLMs. Oikos, 130(8), 1335-1345.

Edwards, Collin; Crone, Elizxabeth (2021). Estimating abundance and phenology from transect count data with GLMs [Dataset]. Dryad. https://doi.org/10.5061/dryad.rn8pk0p9h
}
\usage{
ButterflyCounts
}
\description{
Counts of the baltimore checkerspot butterfly from 2012-2020.
}
\keyword{datasets}
