% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{copepod}
\alias{copepod}
\title{copepod}
\format{
A data frame with 120 rows and 2 variables:
\describe{
\item{\code{Temp}}{double Ambient water temperature (C), ranging from 5C to 30C.}
\item{\code{Oxygen}}{double Mass-specific oxygen consumption rate (\eqn{\mu g} \eqn{O_2} / mg / hour).}
}
}
\source{
Data were simulated to represent an often observed non-linear relationship between temperature and metabolic rates.
}
\usage{
copepod
}
\description{
Simulated data set containing measurements of copepod oxygen consumption rates at different ambient water temperatures
}
\keyword{datasets}
