% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creel}
\alias{creel}
\title{Hours fished by anglers per fishery sector}
\format{
A data frame with 300 rows and 2 variables:
\describe{
\item{fishery}{Fishery sector (factor: \code{Harvest}, \code{Non.Tournament},
\code{Tournament}).}
\item{hours}{Reported hours fished in the year of interest.}
}
}
\source{
Simulated by B. Staton,
\url{https://github.com/bstaton1/au-r-workshop-data}.
}
\usage{
creel
}
\description{
Number of hours fished by 300 anglers, by the fishery sector each angler
participated in.
}
\keyword{datasets}
