% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_catch}
\alias{daily_catch}
\title{Daily commercial harvest of pink salmon}
\format{
A data frame with 50 rows and 51 variables: \code{doy}, the day of the
year (160--209), and 50 columns \code{y_1917}, \code{y_1919}, ..., \code{y_2015} with
the catch on that day in the given year.
}
\source{
Simulated by B. Staton,
\url{https://github.com/bstaton1/au-r-workshop-data}.
}
\usage{
daily_catch
}
\description{
Daily harvest of pink salmon (\emph{Oncorhynchus gorbuscha}) in a commercial
fishery, for the same 50 days in each of 50 odd-numbered years
(1917--2015). Wide format.
}
\seealso{
\link{daily_escape}
}
\keyword{datasets}
