% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feeding}
\alias{feeding}
\title{Predator feeding experiment}
\format{
A data frame with 50 rows and 2 variables:
\describe{
\item{prey}{Prey density.}
\item{cons}{Number of prey consumed.}
}
}
\source{
Simulated by B. Staton,
\url{https://github.com/bstaton1/au-r-workshop-data}.
}
\usage{
feeding
}
\description{
Results of an experiment in which prey density was manipulated and the
number of prey consumed by a predator in a fixed time interval was
recorded. Replicated 50 times at random prey densities.
}
\keyword{datasets}
