% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth}
\alias{growth}
\title{Fish age and length}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{age}{Age (years).}
\item{length}{Total length (mm).}
}
}
\source{
Simulated by B. Staton,
\url{https://github.com/bstaton1/au-r-workshop-data}.
}
\usage{
growth
}
\description{
100 fish sampled at random from a population, with age and length
recorded. Used for the von Bertalanffy bootstrap example.
}
\keyword{datasets}
