% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kelp}
\alias{Kelp}
\title{Kelp}
\format{
A data frame with 38 rows and 5 variables:
\describe{
\item{\code{Response}}{double shallow (upper) distributional limit of the subtidal kelp \emph{Macrocystis pyrifera}}
\item{\code{OD}}{double wave orbital displacement (in meters)}
\item{\code{BD}}{double wave breaking depth (in meters)}
\item{\code{LTD}}{double average tidal height (in meters)}
\item{\code{W}}{double wind velocity (in meters/s)}
}
}
\source{
Graham, M. H. (2003). Confronting multicollinearity in ecological multiple regression. Ecology, 84(11), 2809-2815. \url{https://esapubs.org/archive/ecol/E084/073/suppl-1.htm}
}
\usage{
Kelp
}
\description{
Data from a study of the effect of various environmental factors (wave orbital displacement, wave breaking depth,
wind velocity, and mean tidal height) on the shallow (upper) distributional limit of the subtidal kelp \emph{Macrocystis pyrifera}.
}
\keyword{datasets}
