% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ponds}
\alias{ponds}
\title{Nutrient addition pond experiment}
\format{
A data frame with 20 rows and 8 variables:
\describe{
\item{pond}{Pond identifier (factor: \code{S.28}, \code{S.30}).}
\item{treatment}{Treatment (factor: \code{Control}, \code{Add}).}
\item{replicate}{Replicate within pond and treatment.}
\item{chl.a}{Average chlorophyll \emph{a} in the mesocosm.}
\item{daph, bosm, cope, chao}{Average density of four zooplankton taxa
(\emph{Daphnia}, \emph{Bosmina}, copepods, \emph{Chaoborus}).}
}
}
\source{
Simulated by B. Staton,
\url{https://github.com/bstaton1/au-r-workshop-data}.
}
\usage{
ponds
}
\description{
A hypothetical mesocosm experiment. Nutrients were added to 5 replicate
mesocosms in each of two ponds, with 5 control mesocosms per pond.
}
\details{
The original CSV file contains a deliberate spreadsheet error
(\verb{#VALUE!} in \code{chao}); in this data frame it has been converted to \code{NA}.
The unmodified file is available with
\code{system.file("extdata", "ponds.csv", package = "qsbsdata")}.
}
\keyword{datasets}
