% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sockeye_redfish}
\alias{sockeye_redfish}
\title{Redfish Lake sockeye salmon biology}
\format{
A data frame with 44 rows and 6 variables:
\describe{
\item{year}{Year of the observation.}
\item{type}{Origin of the fish (factor: \code{hatch}, \code{wild}).}
\item{weight}{Average weight (g).}
\item{fecund}{Average number of eggs carried by females.}
\item{egg_size}{Diameter of the average egg (mm).}
\item{survival}{Average percent survival of eggs to the eyed-egg stage.}
}
}
\source{
Kline, P.A. & Flagg, T.A. (2004), as distributed with
\url{https://github.com/bstaton1/au-r-workshop-data}.
}
\usage{
sockeye_redfish
}
\description{
Biological variables for hatchery and wild Redfish Lake sockeye salmon
(\emph{Oncorhynchus nerka}), Idaho, in paired years.
}
\details{
In Staton's book and in the original file (\code{sockeye.csv}, available via
\code{system.file("extdata", "sockeye.csv", package = "qsbsdata")}) this data
set is called \code{sockeye}. It was renamed here to avoid a clash with
\code{\link{sockeye}}, the Fraser River escapement data from
\pkg{Data4Ecologists}.
}
\keyword{datasets}
