% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streams}
\alias{streams}
\title{Stream width and flow}
\format{
A data frame with 20 rows and 3 variables:
\describe{
\item{state}{State the stream is located in (factor).}
\item{stream_width}{Stream width.}
\item{flow}{Flow rate (volume/s).}
}
}
\source{
Simulated by B. Staton,
\url{https://github.com/bstaton1/au-r-workshop-data}.
}
\usage{
streams
}
\description{
Width and flow of 20 streams in the south-eastern United States.
}
\keyword{datasets}
